+++ src/main/java/hs/kr/backend/devpals/domain/notice/controller/NoticeController.java
@@ -25,29 +25,6 @@ public class NoticeController {
     private final NoticeService noticeService;
 
     @Operation(
-            summary = "공지사항 등록",
-            description = "관리자가 새로운 공지사항을 등록합니다.",
-            responses = {
-                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "공지사항 등록 성공"),
-                    @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                            responseCode = "403",
-                            description = "권한이 없는 경우",
-                            content = @Content(
-                                    mediaType = "application/json",
-                                    schema = @Schema(implementation = ApiResponse.class),
-                                    examples = @ExampleObject(value = "{\"success\": false, \"message\": \"관리자 권한이 필요합니다.\", \"data\": null}")
-                            )
-                    )
-            }
-    )
-    @PostMapping
-    public ResponseEntity<ApiResponse<String>> createNotice(
-            @RequestHeader("Authorization") String token,
-            @RequestBody NoticeDTO noticeDTO) {
-        return noticeService.createNotice(token, noticeDTO);
-    }
-
-    @Operation(
             summary = "공지사항 전체 조회",
             description = "등록된 모든 공지사항을 최신순으로 페이징 처리하여 조회합니다.\n\n" +
                     "- `page`는 0부터 시작합니다.\n" +
@@ -86,52 +63,5 @@ public class NoticeController {
     @GetMapping("/{noticeId}")
     public ResponseEntity<ApiResponse<NoticeDetailResponse>> getNotice(@PathVariable Long noticeId) {
         return noticeService.getNotice(noticeId);
-    }
-
-    @Operation(
-            summary = "공지사항 수정",
-            description = "기존 공지사항을 수정합니다.",
-            responses = {
-                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "공지사항 수정 성공"),
-                    @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                            responseCode = "403",
-                            description = "권한이 없는 경우",
-                            content = @Content(
-                                    mediaType = "application/json",
-                                    schema = @Schema(implementation = ApiResponse.class),
-                                    examples = @ExampleObject(value = "{\"success\": false, \"message\": \"관리자 권한이 필요합니다.\", \"data\": null}")
-                            )
-                    )
-            }
-    )
-    @PutMapping("/{noticeId}")
-    public ResponseEntity<ApiResponse<String>> updateNotice(
-            @RequestHeader("Authorization") String token,
-            @PathVariable Long noticeId,
-            @RequestBody NoticeDTO noticeDTO) {
-        return noticeService.updateNotice(token, noticeId, noticeDTO);
-    }
-
-    @Operation(
-            summary = "공지사항 삭제",
-            description = "특정 공지사항을 삭제합니다.",
-            responses = {
-                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "공지사항 삭제 성공"),
-                    @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                            responseCode = "403",
-                            description = "권한이 없는 경우",
-                            content = @Content(
-                                    mediaType = "application/json",
-                                    schema = @Schema(implementation = ApiResponse.class),
-                                    examples = @ExampleObject(value = "{\"success\": false, \"message\": \"관리자 권한이 필요합니다.\", \"data\": null}")
-                            )
-                    )
-            }
-    )
-    @DeleteMapping("/{noticeId}")
-    public ResponseEntity<ApiResponse<String>> deleteNotice(
-            @RequestHeader("Authorization") String token,
-            @PathVariable Long noticeId) {
-        return noticeService.deleteNotice(token, noticeId);
     }
 }
